[refactor] rename TotalLectureSWListTab to LectureSWListTab
- integrate names
