+++ src/components/tabs/LectureSWListTab/styled.ts
